// frontend/src/api/requests/models/Page_RegulatorySchema_.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { RegulatorySchema } from './RegulatorySchema';

export type Page_RegulatorySchema_ = {
    items: Array<RegulatorySchema>;
    page: (number | null);
    pages?: (number | null);
    size: (number | null);
    total: (number | null);
};